Wasm 3.0

+++ Wacs.Core/Runtime/Types/MemoryInstance.cs
@@ -37,7 +37,7 @@ namespace Wacs.Core.Runtime.Types
             if (type.Limits.Minimum > Constants.HostMaxPages)
                 throw new InstantiationException($"Cannot allocate memory of size {type.Limits.Minimum}");
             
-            uint initialSize = (type.Limits.Minimum)* Constants.PageSize;
+            long initialSize = (type.Limits.Minimum)* Constants.PageSize;
             Data = new byte[initialSize];
         }
 
@@ -51,10 +51,10 @@ namespace Wacs.Core.Runtime.Types
         /// <summary>
         /// @Spec 4.5.3.9. Growing memories
         /// </summary>
-        public bool Grow(uint numPages)
+        public bool Grow(long numPages)
         {
-            uint oldNumPages = (uint)(Data.Length / Constants.PageSize);
-            uint newNumPages = oldNumPages + numPages;
+            long oldNumPages = Data.Length / Constants.PageSize;
+            long newNumPages = oldNumPages + numPages;
 
             if (newNumPages > Constants.HostMaxPages)
                 return false;
